Created error message when email is already in db
When the user type email, which is already in db,
an error message will be displayed on the same page.
For this flash from flask is imported and for
loop in registration.html is used.
Added the need to fill out all forms on login and
registration pages(required=true). Added small
css style features.